adding src/utils/slugs.ts
